requirements.txt
# Python 第三方依赖（由仓库内 .py 文件的 import 汇总）
# PyTorch / CUDA：请按 https://pytorch.org/get-started/locally/ 选择与系统匹配的版本；
# 若已用 conda 安装 pytorch torchvision，可从这里注释掉对应行再用 pip 装其余包。

numpy>=1.24
pandas>=2.0
torch>=2.0
torchvision>=0.15
torch-geometric>=2.3
tqdm>=4.65
matplotlib>=3.7
scikit-learn>=1.3
ujson>=5.8
Pillow>=10.0
requests>=2.31
sentence-transformers>=2.2
faiss-cpu>=1.7.4
